Authorization with Gates
-Added authorization for Events and Attendees.
-Added authentication for attendees in the create and delete methods.

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -8,6 +8,7 @@ use App\Http\Resources\AttendeeResource;
 use App\Http\Traits\CanLoadRelationships;
 use App\Models\Event;
 use App\Models\Attendee;
+use Illuminate\Support\Facades\Gate;
 
 class AttendeeController extends Controller
 {
@@ -62,8 +63,14 @@ class AttendeeController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $event, Attendee $attendee)
+    public function destroy(Event $event, Attendee $attendee)
     {
+        if(Gate::denies('delete-attendee', [$event, $attendee])) {
+            abort(403, 'You are not authorized.');
+        }
+
+        // $this->authorize('delete-attendee');
+
         $attendee->delete();
 
         return response(status: 204);
